requirements.txt
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
line-bot-sdk==1.8.0
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
psycopg2-binary
dj-database-url